% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAIBYE.R
\docType{data}
\name{bhai_rates}
\alias{bhai_rates}
\title{BHAI rates per 100,000 population (Germany vs EU/EEA)}
\format{
A data frame with columns:
\describe{
\item{geo}{Geography ("Germany", "EU/EEA").}
\item{sample}{Data source ("German PPS", "ECDC PPS (EU/EEA)").}
\item{hai}{HAI type: HAP, UTI, BSI, SSI, CDI.}
\item{metric}{One of "HAIs", "Deaths", "DALYs".}
\item{per100k}{Rate per 100{,}000 population.}
\item{per100k_low}{Lower 95\\% UI.}
\item{per100k_high}{Upper 95\\% UI.}
}
}
\source{
Zacher et al. (2019), Eurosurveillance.
}
\usage{
data(bhai_rates)
}
\description{
Population-standardised rates by HAI type and metric (HAIs, Deaths, DALYs)
with 95\\% UI for German PPS and ECDC PPS (EU/EEA).
}
\keyword{datasets}
